requirements.txt
fastapi==0.89.0
uvicorn==0.20.0
sqlalchemy==1.4.46
pydantic~=1.10.4
PyJWT~=2.6.0
Logbook~=1.5.3
starlette~=0.22.0
pymysql==1.0.2
requests==2.28.1